// src/types/pokemon.d.ts
export interface StatClass {
    name: string;
    url:  string;
}

export interface Stat {
    base_stat: number;
    effort:    number;
    stat:      StatClass;
}